Moved app info to settings page + additional cleanup
