+++ src/app/home-page/button-bar/button-bar.tsx
@@ -1,9 +1,7 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import SaveConfigDialog from "./save-config-dialog";
 import LoadConfigDialog from "./load-config-dialog";
-import { getName, getTauriVersion, getVersion } from "@tauri-apps/api/app";
-import { invoke } from "@tauri-apps/api/tauri";
-import { useConfigStore } from "../../stores/config-store";
+
 
 export default function ButtonBar({ handleGenerateHDRImage }: any) {
 
@@ -11,31 +9,6 @@ export default function ButtonBar({ handleGenerateHDRImage }: any) {
     useState<boolean>(false);
   const [showLoadConfigDialog, setShowLoadConfigDialog] =
     useState<boolean>(false);
-
-  const [appVersion, setAppVersion] = useState<string>("");
-  const [appName, setAppName] = useState<string>("");
-  const [tauriVersion, setTauriVersion] = useState<string>("");
-
-  const [savedConfigs, setSavedConfigs] = useState<[]>([]);
-
-  // Loads the saved configurations from app config dir using a backend command
-  async function getSavedConfigs() {
-    const json: string = await invoke("get_saved_configs");
-    const configs = JSON.parse(json).configurations;
-    setSavedConfigs(configs);
-  }
-
-  useEffect(() => {
-    // Retrieves app name, app version, and tauri version from Tauri API
-    async function fetchAppInfo() {
-      setAppVersion(await getVersion());
-      setAppName(await getName());
-      setTauriVersion(await getTauriVersion());
-    }
-
-    getSavedConfigs();
-    fetchAppInfo();
-  }, []);
 
   return (
     <div className="fixed bottom-0 left-0 w-full bg-gray-300 flex justify-around py-4">
@@ -46,23 +19,19 @@ export default function ButtonBar({ handleGenerateHDRImage }: any) {
         Load Configuration
       </button>
       <button
-        onClick={() => setShowSaveConfigDialog((prev: any) => !prev)}
+        onClick={() => setShowSaveConfigDialog(!showSaveConfigDialog)}
         className="w-max bg-gray-600 hover:bg-gray-500 text-gray-300 font-semibold py-1 px-4 border-gray-400 rounded"
       >
         Save Configuration
       </button>
       {showSaveConfigDialog && (
         <SaveConfigDialog
-          savedConfigs={savedConfigs}
-          setSavedConfigs={setSavedConfigs}
           toggleDialog={() => setShowSaveConfigDialog(!showSaveConfigDialog)}
         />
       )}
 
       {showLoadConfigDialog && (
         <LoadConfigDialog
-          savedConfigs={savedConfigs}
-          getSavedConfigs={getSavedConfigs}
           toggleDialog={() => setShowLoadConfigDialog(!showLoadConfigDialog)}
         />
       )}
